Fix type mismatch and suppress warnings - Battery temp Float to Double conversion, suppress unused DateHeaderRow and assigned value warnings

+++ java/com/mobeetest/worker/ui/activities/main/pages/composables/device/BatterySectionFields.kt
@@ -73,7 +73,7 @@ fun BatterySectionFields(
             index = index++,
             iconRes = iconRes,
             label = "Temperature",
-            temperatureC = b.temperatureC,
+            temperatureC = b.temperatureC.toDouble(),
             infoDescription = "Current battery temperature in degrees Celsius."
         )
     } else {
